+++ src/controllers/baseConhecimento.go
@@ -2,11 +2,8 @@ package controllers
 
 import (
 	"app/src/config"
-	"app/src/cookies"
-	"app/src/modelos"
 	"app/src/requisicoes"
 	"app/src/respostas"
-	"app/src/utils"
 	"bytes"
 	"encoding/json"
 	"fmt"
@@ -15,307 +12,6 @@ import (
 
 	"github.com/gorilla/mux"
 )
-
-//Carrega tela do formulario
-func CarregarTelaDeCriarPublicacao(w http.ResponseWriter, r *http.Request) {
-
-	url2 := fmt.Sprintf("%s/clientes", config.APIURL)
-	response2, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, url2, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-	if response2.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, response2)
-		return
-	}
-	var cliente []modelos.Cliente
-	if erro = json.NewDecoder(response2.Body).Decode(&cliente); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	url3 := fmt.Sprintf("%s/categorias", config.APIURL)
-	response3, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, url3, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-	if response3.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, response2)
-		return
-	}
-	var categoria []modelos.Post_Categoria
-	if erro = json.NewDecoder(response3.Body).Decode(&categoria); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	url4 := fmt.Sprintf("%s/sites", config.APIURL)
-	response4, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, url4, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-	if response4.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, response2)
-		return
-	}
-	var site []modelos.Site
-	if erro = json.NewDecoder(response4.Body).Decode(&site); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	cookie, _ := cookies.Ler(r)
-	usuarioID, _ := strconv.ParseUint(cookie["id"], 10, 64)
-	nomeUsuario, _ := cookie["nome"]
-	login_nt, _ := cookie["login_nt"]
-	re, _ := cookie["re"]
-	cargo, _ := cookie["cargo"]
-	v_homeoffice, _ := cookie["v_homeoffice"]
-	v_homeoffice_chamados, _ := cookie["v_homeoffice_chamados"]
-	v_homeoffice_chamados_mudar_analista, _ := cookie["v_homeoffice_chamados_mudar_analista"]
-	v_usuarios, _ := cookie["v_usuarios"]
-	v_impressoras, _ := cookie["v_impressoras"]
-	v_bdc_posts, _ := cookie["v_bdc_posts"]
-	v_bdc_adm, _ := cookie["v_bdc_adm"]
-	v_imdb, _ := cookie["v_imdb"]
-	v_gsa, _ := cookie["v_gsa"]
-	v_catraca, _ := cookie["v_catraca"]
-	v_bh, _ := cookie["v_bh"]
-	v_mapa_operacional, _ := cookie["v_mapa_operacional"]
-	site_nome, _ := cookie["Site"]
-
-	utils.ExecutarTemplate(w, "criar-publicacao.html", struct {
-		Cliente                              []modelos.Cliente
-		Categoria                            []modelos.Post_Categoria
-		Site                                 []modelos.Site
-		UsuarioID                            uint64
-		Nome                                 string
-		Login_NT                             string
-		RE                                   string
-		Cargo                                string
-		V_HOMEOFFICE                         string
-		V_HOMEOFFICE_CHAMADOS                string
-		V_HOMEOFFICE_CHAMADOS_MUDAR_ANALISTA string
-		V_USUARIOS                           string
-		V_IMPRESSORAS                        string
-		V_BDC_POSTS                          string
-		V_BDC_ADM                            string
-		V_IMDB                               string
-		V_GSA                                string
-		V_CATRACA                            string
-		V_BH                                 string
-		V_MAPA_OPERACIONAL                   string
-		SiteNome                             string
-		Pagina                               string
-	}{
-
-		Cliente:                              cliente,
-		Categoria:                            categoria,
-		Site:                                 site,
-		UsuarioID:                            usuarioID,
-		Nome:                                 nomeUsuario,
-		Login_NT:                             login_nt,
-		RE:                                   re,
-		Cargo:                                cargo,
-		V_HOMEOFFICE:                         v_homeoffice,
-		V_HOMEOFFICE_CHAMADOS:                v_homeoffice_chamados,
-		V_HOMEOFFICE_CHAMADOS_MUDAR_ANALISTA: v_homeoffice_chamados_mudar_analista,
-		V_USUARIOS:                           v_usuarios,
-		V_IMPRESSORAS:                        v_impressoras,
-		V_BDC_POSTS:                          v_bdc_posts,
-		V_BDC_ADM:                            v_bdc_adm,
-		V_IMDB:                               v_imdb,
-		V_GSA:                                v_gsa,
-		V_CATRACA:                            v_catraca,
-		V_BH:                                 v_bh,
-		V_MAPA_OPERACIONAL:                   v_mapa_operacional,
-		SiteNome:                             site_nome,
-		Pagina:                               "Mapa Operações",
-	})
-}
-
-//Carrega página principal
-func CarregarPaginaDePublicacoes(w http.ResponseWriter, r *http.Request) {
-	url := fmt.Sprintf("%s/docs", config.APIURL)
-	response, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, url, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	if response.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, response)
-		return
-	}
-	var publicacoes []modelos.Post
-	if erro = json.NewDecoder(response.Body).Decode(&publicacoes); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	//DXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
-
-	urlProcedimentos := fmt.Sprintf("%s/docs-procedimentos", config.APIURL)
-	urlTopologia := fmt.Sprintf("%s/docs-topologia", config.APIURL)
-	urlHomeoffice := fmt.Sprintf("%s/docs-homeoffice", config.APIURL)
-	urlDocumentos := fmt.Sprintf("%s/docs-documentos", config.APIURL)
-	urlDiversos := fmt.Sprintf("%s/docs-diversos", config.APIURL)
-
-	//Procedimentos
-	responseProcedimentos, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, urlProcedimentos, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	if responseProcedimentos.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, responseProcedimentos)
-		return
-	}
-	var publicacoesProcedimentos []modelos.Post
-	if erro = json.NewDecoder(responseProcedimentos.Body).Decode(&publicacoesProcedimentos); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-	//Topologia
-
-	responseTopologia, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, urlTopologia, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	if responseTopologia.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, responseTopologia)
-		return
-	}
-	var publicacoesTopologia []modelos.Post
-	if erro = json.NewDecoder(responseTopologia.Body).Decode(&publicacoesTopologia); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-	//Homeoffice
-
-	responseHomeoffice, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, urlHomeoffice, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	if responseHomeoffice.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, responseHomeoffice)
-		return
-	}
-	var publicacoesHomeoffice []modelos.Post
-	if erro = json.NewDecoder(responseHomeoffice.Body).Decode(&publicacoesHomeoffice); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-	//
-
-	responseDocumentos, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, urlDocumentos, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	if responseDocumentos.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, responseDocumentos)
-		return
-	}
-	var publicacoesDocumentos []modelos.Post
-	if erro = json.NewDecoder(responseDocumentos.Body).Decode(&publicacoesDocumentos); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	//Diversos
-
-	responseDiversos, erro := requisicoes.FazerRequisicaoComAutenticacao(r, http.MethodGet, urlDiversos, nil)
-	if erro != nil {
-		respostas.JSON(w, http.StatusInternalServerError, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	if responseDiversos.StatusCode >= 400 {
-		respostas.TratarStatusCodeDeErro(w, responseDiversos)
-		return
-	}
-	var publicacoesDiversos []modelos.Post
-	if erro = json.NewDecoder(responseDiversos.Body).Decode(&publicacoesDiversos); erro != nil {
-		respostas.JSON(w, http.StatusUnprocessableEntity, respostas.ErroAPI{Erro: erro.Error()})
-		return
-	}
-
-	cookie, _ := cookies.Ler(r)
-	usuarioID, _ := strconv.ParseUint(cookie["id"], 10, 64)
-	nomeUsuario, _ := cookie["nome"]
-	login_nt, _ := cookie["login_nt"]
-	re, _ := cookie["re"]
-	cargo, _ := cookie["cargo"]
-	v_homeoffice, _ := cookie["v_homeoffice"]
-	v_homeoffice_chamados, _ := cookie["v_homeoffice_chamados"]
-	v_homeoffice_chamados_mudar_analista, _ := cookie["v_homeoffice_chamados_mudar_analista"]
-	v_usuarios, _ := cookie["v_usuarios"]
-	v_impressoras, _ := cookie["v_impressoras"]
-	v_bdc_posts, _ := cookie["v_bdc_posts"]
-	v_bdc_adm, _ := cookie["v_bdc_adm"]
-	v_imdb, _ := cookie["v_imdb"]
-	v_gsa, _ := cookie["v_gsa"]
-	v_catraca, _ := cookie["v_catraca"]
-	v_bh, _ := cookie["v_bh"]
-	v_mapa_operacional, _ := cookie["v_mapa_operacional"]
-	site_nome, _ := cookie["Site"]
-
-	utils.ExecutarTemplate(w, "docs.html", struct {
-		Publicacoes                          []modelos.Post
-		UsuarioID                            uint64
-		Nome                                 string
-		Login_NT                             string
-		RE                                   string
-		Cargo                                string
-		V_HOMEOFFICE                         string
-		V_HOMEOFFICE_CHAMADOS                string
-		V_HOMEOFFICE_CHAMADOS_MUDAR_ANALISTA string
-		V_USUARIOS                           string
-		V_IMPRESSORAS                        string
-		V_BDC_POSTS                          string
-		V_BDC_ADM                            string
-		V_IMDB                               string
-		V_GSA                                string
-		V_CATRACA                            string
-		V_BH                                 string
-		V_MAPA_OPERACIONAL                   string
-		SiteNome                             string
-		Pagina                               string
-	}{
-		Publicacoes: publicacoes,
-
-		UsuarioID:                            usuarioID,
-		Nome:                                 nomeUsuario,
-		Login_NT:                             login_nt,
-		RE:                                   re,
-		Cargo:                                cargo,
-		V_HOMEOFFICE:                         v_homeoffice,
-		V_HOMEOFFICE_CHAMADOS:                v_homeoffice_chamados,
-		V_HOMEOFFICE_CHAMADOS_MUDAR_ANALISTA: v_homeoffice_chamados_mudar_analista,
-		V_USUARIOS:                           v_usuarios,
-		V_IMPRESSORAS:                        v_impressoras,
-		V_BDC_POSTS:                          v_bdc_posts,
-		V_BDC_ADM:                            v_bdc_adm,
-		V_IMDB:                               v_imdb,
-		V_GSA:                                v_gsa,
-		V_CATRACA:                            v_catraca,
-		V_BH:                                 v_bh,
-		V_MAPA_OPERACIONAL:                   v_mapa_operacional,
-		SiteNome:                             site_nome,
-		Pagina:                               "Mapa Operações",
-		//Pagina:    "Chamado: " + chamado.Chamado,
-	})
-}
 
 //chama a api para cadastrar o usuario no db
 func CriarPublicacao(w http.ResponseWriter, r *http.Request) {
